chore: decommission SAML support for Singpass / Corppass (#5469)
* chore: remove @opengovsg/spcp-auth-client from package.json

* chore: teardown eservice id validation route

* chore: teardown old spcp login routes

* chore: teardown spcp-auth-client

* chore: remove unused getRedirectTarget

* chore: teardown unused exports

* chore: teardown unused env vars, update docs

* chore: remove unused import
