+++ __tests__/e2e/setup/.test-env
@@ -1,14 +1,6 @@
-SP_FORMSG_KEY_PATH=./node_modules/@opengovsg/mockpass/static/certs/key.pem
-SP_FORMSG_CERT_PATH=./node_modules/@opengovsg/mockpass/static/certs/server.crt
-SP_IDP_CERT_PATH=./node_modules/@opengovsg/mockpass/static/certs/spcp.crt
-
-SINGPASS_ASSERT_ENDPOINT=http://localhost:5000/singpass/login
-SINGPASS_IDP_LOGIN_URL=http://localhost:5156/singpass/logininitial
-SINGPASS_IDP_ENDPOINT=http://localhost:5156/singpass/soap
-SINGPASS_PARTNER_ENTITY_ID=http://staging-dsd.form.sg/forms/soap
-SINGPASS_IDP_ID=https://saml-internet.singpass.gov.sg/FIM/sps/SingpassIDPFed/saml20
 MOCKPASS_PORT=5156
-SINGPASS_ESRVC_ID=Test-eServiceId-Sp
+#Needed for MyInfo
+SINGPASS_ESRVC_ID=Test-eServiceId-Sp 
 
 SP_OIDC_NDI_DISCOVERY_ENDPOINT=https://stg-id.singpass.gov.sg/.well-known/openid-configuration
 SP_OIDC_NDI_JWKS_ENDPOINT=https://stg-id.singpass.gov.sg/.well-known/keys
