+++ docs/DEPLOYMENT_SETUP.md
@@ -297,11 +297,7 @@ Note that MyInfo is currently not supported for storage mode forms and enabling 
 | Variable                         | Description                                                                                                                                                                                     |
 | :------------------------------- | ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
 | `SPCP_COOKIE_MAX_AGE_PRESERVED`  | Duration of SingPass JWT before expiry in milliseconds. Defaults to 30 days.                                                                                                                    |
-| `SINGPASS_ESRVC_ID`              | e-service ID registered with National Digital Identity office for SingPass authentication.                                                                                                      |
-| `SINGPASS_PARTNER_ENTITY_ID`     | Partner ID registered with National Digital Identity Office for SingPass authentication.                                                                                                        |
-| `SINGPASS_IDP_LOGIN_URL`         | URL of SingPass Login Page.                                                                                                                                                                     |
-| `SINGPASS_IDP_ENDPOINT`          | URL to retrieve NRIC of SingPass-validated user from.                                                                                                                                           |
-| `SINGPASS_IDP_ID`                | Partner ID of National Digital Identity Office for SingPass authentication.                                                                                                                     |
+| `SINGPASS_ESRVC_ID`              | e-service ID registered with National Digital Identity office for SingPass authentication. Needed for MyInfo.                                                                                   |
 | `SP_OIDC_NDI_DISCOVERY_ENDPOINT` | NDI's Singpass OIDC Discovery Endpoint                                                                                                                                                          |
 | `SP_OIDC_NDI_JWKS_ENDPOINT`      | NDI's Singpass OIDC JWKS Endpoint                                                                                                                                                               |
 | `SP_OIDC_RP_CLIENT_ID`           | The Relying Party's Singpass Client ID as registered with NDI                                                                                                                                   |
@@ -314,17 +310,6 @@ Note that MyInfo is currently not supported for storage mode forms and enabling 
 | `CP_OIDC_RP_REDIRECT_URL`        | The Relying Party's Corppass Redirect URL                                                                                                                                                       |
 | `CP_OIDC_RP_JWKS_PUBLIC_PATH`    | Path to the Relying Party's Public Json Web Key Set used for Corppass-related communication with NDI. This will be hosted at api/v3/corppass/.well-known/jwks.json endpoint.                    |
 | `CP_OIDC_RP_JWKS_SECRET_PATH`    | Path to the Relying Party's Secret Json Web Key Set used for Corppass-related communication with NDI                                                                                            |
-| `CORPPASS_ESRVC_ID`              | e-service ID registered with National Digital Identity office for CorpPass authentication.                                                                                                      |
-| `CORPPASS_PARTNER_ENTITY_ID`     | Partner ID registered with National Digital Identity Office for CorpPass authentication.                                                                                                        |
-| `CORPPASS_IDP_LOGIN_URL`         | URL of CorpPass Login Page.                                                                                                                                                                     |
-| `CORPPASS_IDP_ENDPOINT`          | URL to retrieve UEN of CorpPass-validated user from.                                                                                                                                            |
-| `CORPPASS_IDP_ID`                | Partner ID of National Digital Identity Office for CorpPass authentication.                                                                                                                     |
-| `SP_FORMSG_KEY_PATH`             | Path to X.509 key used for SingPass related communication with National Digital Identity office.                                                                                                |
-| `SP_FORMSG_CERT_PATH`            | Path to X.509 cert used for SingPass related communication with National Digital Identity office.                                                                                               |
-| `SP_IDP_CERT_PATH`               | Path to National Digital Identity office's X.509 cert used for SingPass related communication.                                                                                                  |
-| `CP_FORMSG_KEY_PATH`             | Path to X.509 key used for CorpPass related communication with National Digital Identity office.                                                                                                |
-| `CP_FORMSG_CERT_PATH`            | Path to X.509 cert used for CorpPass related communication with National Digital Identity office.                                                                                               |
-| `CP_IDP_CERT_PATH`               | Path to National Digital Identity office's X.509 cert used for CorpPass related communication.                                                                                                  |
 | `MYINFO_CLIENT_CONFIG`           | Configures [MyInfoGovClient](https://github.com/opengovsg/myinfo-gov-client). Set this to either`stg` or `prod` to fetch MyInfo data from the corresponding endpoints.                          |
 | `MYINFO_FORMSG_KEY_PATH`         | Filepath to MyInfo private key, which is used to decrypt data and sign requests when communicating with MyInfo.                                                                                 |
 | `MYINFO_CERT_PATH`               | Path to MyInfo's public certificate, which is used to verify their signature.                                                                                                                   |
