+++ package.json
@@ -73,7 +73,6 @@
     "@opengovsg/myinfo-gov-client": "^4.0.1",
     "@opengovsg/ng-file-upload": "^12.2.15",
     "@opengovsg/sgid-client": "1.0.4",
-    "@opengovsg/spcp-auth-client": "^1.4.20",
     "@sentry/browser": "^7.20.1",
     "@sentry/integrations": "^6.19.7",
     "@stablelib/base64": "^1.0.1",
