+++ shared/types/form/form_auth.ts
@@ -1,7 +1,3 @@
 export type PublicFormAuthRedirectDto = { redirectURL: string }
 
-export type PublicFormAuthValidateEsrvcIdDto =
-  | { isValid: true }
-  | { isValid: false; errorCode: string | null }
-
 export type PublicFormAuthLogoutDto = { message: string }
