+++ src/app/modules/form/public-form/__tests__/public-form.controller.spec.ts
@@ -24,7 +24,6 @@ import { MOCK_LOGIN_DOC } from 'src/app/modules/spcp/__tests__/spcp.test.constan
 import { JwtPayload, SpcpForm } from 'src/app/modules/spcp/spcp.types'
 import {
   IFormDocument,
-  IFormSchema,
   IPopulatedForm,
   IPopulatedUser,
   PublicForm,
@@ -43,13 +42,10 @@ import { MyInfoService } from '../../../myinfo/myinfo.service'
 import { SGID_COOKIE_NAME } from '../../../sgid/sgid.constants'
 import {
   CreateRedirectUrlError,
-  FetchLoginPageError,
-  LoginPageValidationError,
   MissingJwtError,
 } from '../../../spcp/spcp.errors'
 import { CpOidcServiceClass } from '../../../spcp/spcp.oidc.service/spcp.oidc.service.cp'
 import { SpOidcServiceClass } from '../../../spcp/spcp.oidc.service/spcp.oidc.service.sp'
-import { SpcpService } from '../../../spcp/spcp.service'
 import { JwtName } from '../../../spcp/spcp.types'
 import {
   AuthTypeMismatchError,
@@ -65,7 +61,6 @@ import { Metatags } from '../public-form.types'
 jest.mock('../public-form.service')
 jest.mock('../../form.service')
 jest.mock('../../../auth/auth.service')
-jest.mock('../../../spcp/spcp.service')
 jest.mock('../../../spcp/spcp.oidc.service/spcp.oidc.service.sp')
 jest.mock('../../../spcp/spcp.oidc.service/spcp.oidc.service.cp')
 jest.mock('../../../myinfo/myinfo.service')
@@ -74,7 +69,6 @@ jest.mock('../../../billing/billing.service')
 const MockFormService = mocked(FormService)
 const MockPublicFormService = mocked(PublicFormService)
 const MockAuthService = mocked(AuthService)
-const MockSpcpService = mocked(SpcpService, true)
 
 const MockMyInfoService = mocked(MyInfoService, true)
 const MockBillingService = mocked(BillingService, true)
@@ -1507,337 +1501,6 @@ describe('public-form.controller', () => {
       expect(mockRes.json).toHaveBeenCalledWith({
         message: 'Successfully logged out.',
       })
-    })
-  })
-
-  describe('handleValidateFormEsrvcId', () => {
-    const MOCK_REQ = expressHandler.mockRequest({
-      params: {
-        formId: new ObjectId().toHexString(),
-      },
-    })
-    const MOCK_REDIRECT_URL = 'www.mockata.com'
-
-    it('should return 200 with isValid set to true when a valid MyInfo form authenticates successfully', async () => {
-      // Arrange
-      const MOCK_FORM = {
-        authType: FormAuthType.MyInfo,
-        esrvcId: '12345',
-      } as MyInfoForm<IFormSchema>
-      const mockRes = expressHandler.mockResponse()
-      const expectedResBody = { isValid: true as const }
-      MockFormService.retrieveFormById.mockReturnValueOnce(okAsync(MOCK_FORM))
-      MockSpcpService.createRedirectUrl.mockReturnValueOnce(
-        ok(MOCK_REDIRECT_URL),
-      )
-      MockSpcpService.fetchLoginPage.mockReturnValueOnce(
-        okAsync('this is raw html'),
-      )
-      MockSpcpService.validateLoginPage.mockReturnValueOnce(ok(expectedResBody))
-
-      // Act
-      await PublicFormController.handleValidateFormEsrvcId(
-        MOCK_REQ,
-        mockRes,
-        jest.fn(),
-      )
-
-      // Assert
-      expect(mockRes.status).toHaveBeenCalledWith(200)
-      expect(mockRes.json).toHaveBeenCalledWith(expectedResBody)
-    })
-
-    it('should return 200 with isValid set to true when a valid SP form authenticates successfully', async () => {
-      // Arrange
-      const MOCK_FORM = {
-        authType: FormAuthType.SP,
-        esrvcId: '12345',
-      } as SpcpForm<IFormSchema>
-      const mockRes = expressHandler.mockResponse()
-      const expectedResBody = { isValid: true as const }
-      MockFormService.retrieveFormById.mockReturnValueOnce(okAsync(MOCK_FORM))
-      MockSpcpService.createRedirectUrl.mockReturnValueOnce(
-        ok(MOCK_REDIRECT_URL),
-      )
-      MockSpcpService.fetchLoginPage.mockReturnValueOnce(
-        okAsync('this is raw html'),
-      )
-      MockSpcpService.validateLoginPage.mockReturnValueOnce(ok(expectedResBody))
-
-      // Act
-      await PublicFormController.handleValidateFormEsrvcId(
-        MOCK_REQ,
-        mockRes,
-        jest.fn(),
-      )
-
-      // Assert
-      expect(mockRes.status).toHaveBeenCalledWith(200)
-      expect(mockRes.json).toHaveBeenCalledWith(expectedResBody)
-    })
-    it('should return 200 with isValid set to false when an invalid MyInfo form authenticates successfully', async () => {
-      // Arrange
-      const MOCK_FORM = {
-        authType: FormAuthType.MyInfo,
-        esrvcId: '12345',
-      } as MyInfoForm<IFormSchema>
-      const mockRes = expressHandler.mockResponse()
-      const expectedResBody = { isValid: false, errorCode: '138' }
-      MockFormService.retrieveFormById.mockReturnValueOnce(okAsync(MOCK_FORM))
-      MockSpcpService.createRedirectUrl.mockReturnValueOnce(
-        ok(MOCK_REDIRECT_URL),
-      )
-      MockSpcpService.fetchLoginPage.mockReturnValueOnce(
-        okAsync('this is raw html'),
-      )
-      MockSpcpService.validateLoginPage.mockReturnValueOnce(ok(expectedResBody))
-
-      // Act
-      await PublicFormController.handleValidateFormEsrvcId(
-        MOCK_REQ,
-        mockRes,
-        jest.fn(),
-      )
-
-      // Assert
-      expect(mockRes.status).toHaveBeenCalledWith(200)
-      expect(mockRes.json).toHaveBeenCalledWith(expectedResBody)
-    })
-
-    it('should return 200 with isValid set to false when an invalid SP form authenticates successfully', async () => {
-      // Arrange
-      const MOCK_FORM = {
-        authType: FormAuthType.SP,
-        esrvcId: '12345',
-      } as SpcpForm<IFormSchema>
-      const mockRes = expressHandler.mockResponse()
-      const expectedResBody = { isValid: false, errorCode: '138' }
-      MockFormService.retrieveFormById.mockReturnValueOnce(okAsync(MOCK_FORM))
-      MockSpcpService.createRedirectUrl.mockReturnValueOnce(
-        ok(MOCK_REDIRECT_URL),
-      )
-      MockSpcpService.fetchLoginPage.mockReturnValueOnce(
-        okAsync('this is raw html'),
-      )
-      MockSpcpService.validateLoginPage.mockReturnValueOnce(ok(expectedResBody))
-
-      // Act
-      await PublicFormController.handleValidateFormEsrvcId(
-        MOCK_REQ,
-        mockRes,
-        jest.fn(),
-      )
-
-      // Assert
-      expect(mockRes.status).toHaveBeenCalledWith(200)
-      expect(mockRes.json).toHaveBeenCalledWith(expectedResBody)
-    })
-
-    it('should return 400 when the form has authType.NIL', async () => {
-      // Arrange
-      const MOCK_FORM = {
-        authType: FormAuthType.NIL,
-        esrvcId: '12345',
-      } as IFormSchema
-      const mockRes = expressHandler.mockResponse()
-      const expectedError = {
-        message:
-          'Please ensure that the form has authentication enabled. Please refresh and try again.',
-      }
-      MockFormService.retrieveFormById.mockReturnValueOnce(okAsync(MOCK_FORM))
-
-      // Act
-      await PublicFormController.handleValidateFormEsrvcId(
-        MOCK_REQ,
-        mockRes,
-        jest.fn(),
-      )
-
-      // Assert
-      expect(mockRes.status).toHaveBeenCalledWith(400)
-      expect(mockRes.json).toHaveBeenCalledWith(expectedError)
-    })
-
-    it('should return 400 when the form has authType.CP', async () => {
-      // Arrange
-      const MOCK_FORM = {
-        authType: FormAuthType.CP,
-        esrvcId: '12345',
-      } as IFormSchema
-      const mockRes = expressHandler.mockResponse()
-      const expectedError = {
-        message:
-          'Please ensure that the form has authentication enabled. Please refresh and try again.',
-      }
-      MockFormService.retrieveFormById.mockReturnValueOnce(okAsync(MOCK_FORM))
-
-      // Act
-      await PublicFormController.handleValidateFormEsrvcId(
-        MOCK_REQ,
-        mockRes,
-        jest.fn(),
-      )
-
-      // Assert
-      expect(mockRes.status).toHaveBeenCalledWith(400)
-      expect(mockRes.json).toHaveBeenCalledWith(expectedError)
-    })
-
-    it('should return 400 when the form does not have an eServiceId', async () => {
-      // Arrange
-      const MOCK_FORM = {
-        authType: FormAuthType.SP,
-      } as SpcpForm<IFormSchema>
-      const mockRes = expressHandler.mockResponse()
-      const expectedError = {
-        message:
-          'This form does not have a valid eServiceId. Please refresh and try again.',
-      }
-      MockFormService.retrieveFormById.mockReturnValueOnce(okAsync(MOCK_FORM))
-
-      // Act
-      await PublicFormController.handleValidateFormEsrvcId(
-        MOCK_REQ,
-        mockRes,
-        jest.fn(),
-      )
-
-      // Assert
-      expect(mockRes.status).toHaveBeenCalledWith(400)
-      expect(mockRes.json).toHaveBeenCalledWith(expectedError)
-    })
-
-    it('should return 404 when there is no form with the given formId', async () => {
-      // Arrange
-      const mockRes = expressHandler.mockResponse()
-      const expectedError = {
-        message:
-          'Could not find the form requested. Please refresh and try again.',
-      }
-      MockFormService.retrieveFormById.mockReturnValueOnce(
-        errAsync(new FormNotFoundError()),
-      )
-
-      // Act
-      await PublicFormController.handleValidateFormEsrvcId(
-        MOCK_REQ,
-        mockRes,
-        jest.fn(),
-      )
-
-      // Assert
-      expect(mockRes.status).toHaveBeenCalledWith(404)
-      expect(mockRes.json).toHaveBeenCalledWith(expectedError)
-    })
-
-    it('should return 500 when the returned html has no title', async () => {
-      // Arrange
-      const MOCK_FORM = {
-        authType: FormAuthType.SP,
-        esrvcId: '12345',
-      } as SpcpForm<IFormSchema>
-      const mockRes = expressHandler.mockResponse()
-      const expectedError = {
-        message: 'Error while contacting SingPass. Please try again.',
-      }
-      MockFormService.retrieveFormById.mockReturnValueOnce(okAsync(MOCK_FORM))
-      MockSpcpService.createRedirectUrl.mockReturnValueOnce(
-        ok(MOCK_REDIRECT_URL),
-      )
-      MockSpcpService.fetchLoginPage.mockReturnValue(okAsync(''))
-      MockSpcpService.validateLoginPage.mockReturnValueOnce(
-        err(new LoginPageValidationError()),
-      )
-
-      // Act
-      await PublicFormController.handleValidateFormEsrvcId(
-        MOCK_REQ,
-        mockRes,
-        jest.fn(),
-      )
-
-      // Assert
-      expect(mockRes.status).toHaveBeenCalledWith(502)
-      expect(mockRes.json).toHaveBeenCalledWith(expectedError)
-    })
-
-    it('should return 500 when a database error occurs', async () => {
-      // Arrange
-      const mockRes = expressHandler.mockResponse()
-      const expectedError = {
-        message: 'Sorry, something went wrong. Please try again.',
-      }
-      MockFormService.retrieveFormById.mockReturnValueOnce(
-        errAsync(new DatabaseError()),
-      )
-
-      // Act
-      await PublicFormController.handleValidateFormEsrvcId(
-        MOCK_REQ,
-        mockRes,
-        jest.fn(),
-      )
-
-      // Assert
-      expect(mockRes.status).toHaveBeenCalledWith(500)
-      expect(mockRes.json).toHaveBeenCalledWith(expectedError)
-    })
-
-    it('should return 500 when the redirect url could not be generated', async () => {
-      // Arrange
-      const MOCK_FORM = {
-        authType: FormAuthType.SP,
-        esrvcId: '12345',
-      } as SpcpForm<IFormSchema>
-      const mockRes = expressHandler.mockResponse()
-      const expectedError = {
-        message: 'Sorry, something went wrong. Please try again.',
-      }
-      MockFormService.retrieveFormById.mockReturnValueOnce(okAsync(MOCK_FORM))
-      MockSpcpService.createRedirectUrl.mockReturnValueOnce(
-        err(new CreateRedirectUrlError()),
-      )
-
-      // Act
-      await PublicFormController.handleValidateFormEsrvcId(
-        MOCK_REQ,
-        mockRes,
-        jest.fn(),
-      )
-
-      // Assert
-      expect(mockRes.status).toHaveBeenCalledWith(500)
-      expect(mockRes.json).toHaveBeenCalledWith(expectedError)
-    })
-
-    it('should return 503 when the login page for the form could not be retrieved', async () => {
-      // Arrange
-      const MOCK_FORM = {
-        authType: FormAuthType.SP,
-        esrvcId: '12345',
-      } as SpcpForm<IFormSchema>
-      const mockRes = expressHandler.mockResponse()
-      const expectedError = {
-        message: 'Failed to contact SingPass. Please try again.',
-      }
-      MockFormService.retrieveFormById.mockReturnValueOnce(okAsync(MOCK_FORM))
-      MockSpcpService.createRedirectUrl.mockReturnValueOnce(
-        ok(MOCK_REDIRECT_URL),
-      )
-      MockSpcpService.fetchLoginPage.mockReturnValue(
-        errAsync(new FetchLoginPageError()),
-      )
-
-      // Act
-      await PublicFormController.handleValidateFormEsrvcId(
-        MOCK_REQ,
-        mockRes,
-        jest.fn(),
-      )
-
-      // Assert
-      expect(mockRes.status).toHaveBeenCalledWith(503)
-      expect(mockRes.json).toHaveBeenCalledWith(expectedError)
     })
   })
 })
