+++ src/app/modules/myinfo/myinfo.routes.ts
@@ -2,7 +2,6 @@ import { Router } from 'express'
 
 import { MYINFO_REDIRECT_PATH } from './myinfo.constants'
 import {
-  handleEServiceIdCheck,
   handleMyInfoLogin,
   handleRedirectURLRequest,
 } from './myinfo.controller'
@@ -15,12 +14,6 @@ export const MyInfoRouter = Router()
  * @deprecated in favour of GET /api/v3/forms/:formId/auth/redirect
  */
 MyInfoRouter.get('/redirect', handleRedirectURLRequest)
-
-/**
- * Serves requests to validate a MyInfo form's SingPass e-service ID.
- * @deprecated in favour of POST /api/v3/forms/:formId/auth/validate
- */
-MyInfoRouter.get('/validate', handleEServiceIdCheck)
 
 /**
  * Serves redirects from MyInfo after user has given consent to provide
