+++ src/app/modules/spcp/spcp.errors.ts
@@ -10,46 +10,10 @@ export class CreateRedirectUrlError extends ApplicationError {
 }
 
 /**
- * Error while fetching SP/CP login page.
- */
-export class FetchLoginPageError extends ApplicationError {
-  constructor(message = 'Error while fetching SP/CP login page') {
-    super(message)
-  }
-}
-
-/**
- * Invalid SP/CP login page.
- */
-export class LoginPageValidationError extends ApplicationError {
-  constructor(message = 'Invalid SP/CP login page') {
-    super(message)
-  }
-}
-
-/**
  * JWT could not be decoded.
  */
 export class VerifyJwtError extends ApplicationError {
   constructor(message = 'Invalid JWT') {
-    super(message)
-  }
-}
-
-/**
- * Invalid OOB params passed to login endpoint.
- */
-export class InvalidOOBParamsError extends ApplicationError {
-  constructor(message = 'Invalid OOB params passed to login endpoint') {
-    super(message)
-  }
-}
-
-/**
- * Error while attempting to retrieve SPCP attributes from SPCP server
- */
-export class RetrieveAttributesError extends ApplicationError {
-  constructor(message = 'Failed to retrieve attributes from SPCP') {
     super(message)
   }
 }
