+++ src/app/modules/spcp/spcp.oidc.service/__tests/spcp.oidc.service.sp.spec.ts
@@ -5,7 +5,6 @@ import { MOCK_COOKIE_AGE } from 'src/app/modules/myinfo/__tests__/myinfo.test.co
 
 import dbHandler from 'tests/unit/backend/helpers/jest-db'
 
-import { spOidcClientConfig } from '../../__tests__/spcp.oidc.test.constants'
 import {
   MOCK_COOKIES,
   MOCK_DECODED_QUERY,
@@ -19,6 +18,7 @@ import {
   MOCK_SP_OIDC_AUTHORISATION_CODE,
   MOCK_SP_OIDC_EXTRACTED_NDI_PAYLOAD,
   MOCK_TARGET,
+  spOidcClientConfig,
 } from '../../__tests__/spcp.test.constants'
 import {
   CreateJwtError,
