+++ src/app/modules/submission/encrypt-submission/encrypt-submission.utils.ts
@@ -35,9 +35,7 @@ import {
 } from '../../sgid/sgid.errors'
 import {
   CreateRedirectUrlError,
-  FetchLoginPageError,
   InvalidJwtError,
-  LoginPageValidationError,
   MissingJwtError,
   VerifyJwtError,
 } from '../../spcp/spcp.errors'
@@ -73,16 +71,6 @@ const errorMapper: MapRouteError = (
       return {
         statusCode: StatusCodes.INTERNAL_SERVER_ERROR,
         errorMessage: coreErrorMessage,
-      }
-    case FetchLoginPageError:
-      return {
-        statusCode: StatusCodes.SERVICE_UNAVAILABLE,
-        errorMessage: 'Failed to contact SingPass. Please try again.',
-      }
-    case LoginPageValidationError:
-      return {
-        statusCode: StatusCodes.BAD_GATEWAY,
-        errorMessage: 'Error while contacting SingPass. Please try again.',
       }
     case SgidMissingJwtError:
     case SgidVerifyJwtError:
