+++ tests/.test-env
@@ -1,12 +1,3 @@
-SP_FORMSG_KEY_PATH=./node_modules/@opengovsg/mockpass/static/certs/key.pem
-SP_FORMSG_CERT_PATH=./node_modules/@opengovsg/mockpass/static/certs/server.crt
-SP_IDP_CERT_PATH=./node_modules/@opengovsg/mockpass/static/certs/spcp.crt
-
-SINGPASS_ASSERT_ENDPOINT=http://localhost:5000/singpass/login
-SINGPASS_IDP_LOGIN_URL=http://localhost:5156/singpass/logininitial
-SINGPASS_IDP_ENDPOINT=http://localhost:5156/singpass/soap
-SINGPASS_PARTNER_ENTITY_ID=http://staging-dsd.form.sg/forms/soap
-SINGPASS_IDP_ID=https://saml-internet.singpass.gov.sg/FIM/sps/SingpassIDPFed/saml20
 MOCKPASS_PORT=5156
 SINGPASS_ESRVC_ID=Test-eServiceId-Sp
 
@@ -29,17 +20,6 @@ MYINFO_CERT_PATH=./node_modules/@opengovsg/mockpass/static/certs/spcp.crt
 MYINFO_CLIENT_ID=mockClientId
 MYINFO_CLIENT_SECRET=mockClientSecret
 MYINFO_JWT_SECRET=mockJwtSecret
-
-CP_FORMSG_KEY_PATH=./node_modules/@opengovsg/mockpass/static/certs/key.pem
-CP_FORMSG_CERT_PATH=./node_modules/@opengovsg/mockpass/static/certs/server.crt
-CP_IDP_CERT_PATH=./node_modules/@opengovsg/mockpass/static/certs/spcp.crt
-
-CORPPASS_ASSERT_ENDPOINT=http://localhost:5000/corppass/login
-CORPPASS_IDP_LOGIN_URL=http://localhost:5156/corppass/logininitial
-CORPPASS_IDP_ENDPOINT=http://localhost:5156/corppass/soap
-CORPPASS_PARTNER_ENTITY_ID=https://staging.form.gov.sg/corppass
-CORPPASS_ESRVC_ID=FORMSG-CP-TEST
-CORPPASS_IDP_ID=https://saml.corppass.gov.sg/FIM/sps/CorpIDPFed/saml20
 
 SHOW_LOGIN_PAGE=true
 IS_SP_MAINTENANCE=Date/Time-SP
